requirements.txt
alembic==1.9.3
APScheduler==3.10.1
Babel==2.11.0
bcrypt==4.0.1
blinker==1.5
cachelib==0.10.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
dnspython==2.3.0
dominate==2.7.0
email-validator==1.3.1
et-xmlfile==1.1.0
Flask==2.2.2
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.0.1
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy==1.5
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
ntplib==0.4.0
oauth2==1.9.0.post1
oauthlib==3.2.2
openpyxl==3.1.2
passlib==1.7.4
Pillow==9.4.0
protobuf==3.20.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-barcode==0.14.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.3
SQLAlchemy-Utils==0.39.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.14
URLObject==2.4.3
visitor==0.1.3
Werkzeug==2.2.2
wincertstore==0.2
WTForms==3.0.1
